ARROW-11563: [Rust] Implement Cast(Utf8, TimeStamp(Nanoseconds, None))
