+++ rust/arrow/src/compute/kernels/mod.rs
@@ -22,6 +22,7 @@ pub mod arithmetic;
 pub mod arity;
 pub mod boolean;
 pub mod cast;
+pub mod cast_utils;
 pub mod comparison;
 pub mod concat;
 pub mod filter;
